pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbiter"
version = "0.1.0"
description = "Access auditing, consistency analysis, blast-radius classification, and trust enforcement for the Pact/Baton stack."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "Jeremy McEntire", email = "jandrewmcentire@gmail.com"}]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
otlp = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "grpcio>=1.60",
]
api = [
    "flask>=3.0",
]
all = [
    "arbiter[otlp,api]",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
arbiter = "arbiter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
